renamed and restructed file structure
